[11.x] Use scout prefix when deleting all indexes (#841)
* use scout prefix when deleting indexes

* document changed behavior

* Update MeilisearchEngine.php

* Update UPGRADE.md

---------

+++ src/Engines/MeilisearchEngine.php
@@ -4,6 +4,7 @@ namespace Laravel\Scout\Engines;
 
 use BackedEnum;
 use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\LazyCollection;
 use Laravel\Scout\Builder;
 use Laravel\Scout\Contracts\UpdatesIndexSettings;
@@ -452,7 +453,9 @@ class MeilisearchEngine extends Engine implements UpdatesIndexSettings
         $indexes = $this->meilisearch->getIndexes($query);
 
         foreach ($indexes->getResults() as $index) {
-            $tasks[] = $index->delete();
+            if (str($index->getUid())->startsWith(Config::get('scout.prefix'))) {
+                $tasks[] = $index->delete();
+            }
         }
 
         return $tasks;
